More optimizations

+++ src/main/java/ranking/IndexedRanking.java
@@ -10,19 +10,19 @@ import java.util.List;
  * Created by janwillem on 01/04/16.
  */
 public class IndexedRanking {
-    private final int[] indices;
+    private final short[] indices;
     private final int maxSize = ConfigFactory.load().getInt("implementation.expectedNumConstraints");
 
     // NB: Assumes that RankedConstraints is sorted & reversed!
     public IndexedRanking(List<RankedConstraint> rankedConstraints) {
-        indices = new int[maxSize];
-        for (int i=0; i < rankedConstraints.size(); i++) {
-            int constraintIndex = rankedConstraints.get(i).getConstraint().getId();
+        indices = new short[maxSize];
+        for (short i=0; i < rankedConstraints.size(); i++) {
+            short constraintIndex = rankedConstraints.get(i).getConstraint().getId();
             indices[constraintIndex] = i;
         }
     }
 
-    public int getIndex(Constraint c) {
+    public short getIndex(Constraint c) {
         return indices[c.getId()];
     }
